+++ dtslint/ts3.5/Array.ts
@@ -83,3 +83,10 @@ pipe(
   us,
   _.spanLeft((u: unknown): u is number => typeof u === 'number')
 )
+
+//
+// lookup
+//
+
+_.lookup(0, [1, 2, 3]) // $ExpectType Option<number>
+_.lookup(0) // $ExpectType <A>(as: A[]) => Option<A>