+++ CHANGELOG.md
@@ -19,8 +19,10 @@ high state of flux, you're at risk of it changing without notice.
 - **Polish**
   - `Array`
     - export `unfold` (@gcanti)
+    - make `lookup` data-last (@gcanti)
   - `ReadonlyArray`
     - export `unfold` (@gcanti)
+    - make `lookup` data-last (@gcanti)
   - `Semigroup`
     - make `fold` data-last (@gcanti)